+++ Stratis.Bitcoin.Features.Wallet/WalletFeePolicy.cs
@@ -2,21 +2,12 @@
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Features.MemoryPool;
 using Stratis.Bitcoin.Features.MemoryPool.Fee;
+using Stratis.Bitcoin.Interfaces;
 using Stratis.Bitcoin.Utilities;
 using System;
 
 namespace Stratis.Bitcoin.Features.Wallet
 {
-    public interface IWalletFeePolicy
-    {
-        void Start();
-        void Stop();
-        Money GetRequiredFee(int txBytes);
-        Money GetMinimumFee(int txBytes, int confirmTarget);
-        Money GetMinimumFee(int txBytes, int confirmTarget, Money targetFee);
-        FeeRate GetFeeRate(int confirmTarget);
-    }
-
     public class WalletFeePolicy : IWalletFeePolicy
     {
         /// <summary>Block policy estimator.</summary>
